requirements.txt
alembic==1.9.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bandit==1.7.4
black==23.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
coverage==7.1.0
flake8==6.0.0
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.10.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
pbr==5.11.1
pkg_resources==0.0.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python-dotenv==0.21.1
PyYAML==6.0
smmap==5.0.0
SQLAlchemy==2.0.0
stevedore==4.1.1
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.0
